requirements.txt
streamlit
elasticsearch
mistralai
stqdm